Lint code, improve UX and implement feedback
Using html-tidy and csslint, the number of warnings have been reduced
and using Prettier the code has been better formatted. A hover animation
has been added to the links at the top of the page as well as an
indication of which page you are currently on. The JavaScript script
tags have now been moved inside the body tag at the end of the page.
Additional details about the project have been added to the README
including more detail about the testing done and a gif of the front page
at the top.