Change H tag order and add to README
Implemented the feedback about the use of H tags, they are now ordered by importance on the site rather than as a proxy for fontsize. Where the text size needs to be different, extra classes have been added to the CSS to override the H tag sizes. Sections of the README have also been expanded on.